pyproject.toml
[tool.poetry]
name = "slackbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.14.3"
python-dotenv = "^0.20.0"
psycopg-pool = "^3.1.1"
psycopg = "^3.0.16"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
